# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  geregetemplateai_internal_http_datatransfers_requests.BPMSaveFormRequest:
    properties:
      name:
        maxLength: 200
        minLength: 1
        type: string
      schema:
        items:
          type: integer
        type: array
    required:
    - name
    - schema
    type: object
...
  geregetemplateai_internal_http_datatransfers_requests.BPMSubmitTaskRequest:
    properties:
      data:
        items:
          type: integer
        type: array
    required:
    - data
    type: object
...
  geregetemplateai_internal_http_datatransfers_responses.BPMEventResponse:
    properties:
      created_at:
        type: string
      detail:
        type: string
      id:
        type: string
      node_id:
        type: string
      type:
        type: string
    type: object
  geregetemplateai_internal_http_datatransfers_responses.BPMFormResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      schema:
        items:
          type: integer
        type: array
      updated_at:
        type: string
    type: object
  geregetemplateai_internal_http_datatransfers_responses.BPMInstanceResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      current_node_id:
        type: string
      definition_id:
        type: string
      id:
        type: string
      status:
        type: string
      variables:
        items:
          type: integer
        type: array
    type: object
...
  geregetemplateai_internal_http_datatransfers_responses.BPMRunResponse:
    properties:
      done:
        type: boolean
      instance:
        $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMInstanceResponse'
      task:
        $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMTaskResponse'
    type: object
  geregetemplateai_internal_http_datatransfers_responses.BPMTaskResponse:
    properties:
      form:
        items:
          type: integer
        type: array
      id:
        type: string
      node_id:
        type: string
      status:
        type: string
    type: object
...
  geregetemplateai_internal_http_handlers_v1.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Gerege Template AI v1.0
    url: https://github.com/snykk/go-rest-boilerplate
  description: Fiber v3 + GORM (PostgreSQL) + Redis дээр суурилсан Clean Architecture
    бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib
    Fikri)-ээс үүсэлтэй; Gin -> Fiber v3 болон sqlx -> GORM руу хөрвүүлсэн.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/snykk/go-rest-boilerplate
  title: Gerege Template AI v1.0 API
  version: "1.0"
paths:
...
  /bpm/forms:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMFormResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Хуваалцсан формуудыг жагсаах
      tags:
      - bpm
    post:
      consumes:
      - application/json
      parameters:
      - description: Form
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.BPMSaveFormRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMFormResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Хуваалцсан форм үүсгэх
      tags:
      - bpm
  /bpm/forms/{id}:
    delete:
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Хуваалцсан форм устгах
      tags:
      - bpm
    put:
      consumes:
      - application/json
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      - description: Form
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.BPMSaveFormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMFormResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Хуваалцсан форм засах (latest-wins)
      tags:
      - bpm
...
  /bpm/instances/{id}/events:
    get:
      parameters:
      - description: Instance ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMEventResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Гүйлтийн audit timeline (events)
      tags:
      - bpm
  /bpm/instances/{id}/task:
    get:
      parameters:
      - description: Instance ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMRunResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Гүйлтийн идэвхтэй даалгаврыг авах (рендерлэх дэлгэц)
      tags:
      - bpm
...
  /bpm/processes/{id}/instances:
    get:
      parameters:
      - description: Process ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMInstanceResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Процессын гүйлтүүдийг (instances) жагсаах
      tags:
      - bpm
  /bpm/processes/{id}/start:
    post:
      parameters:
      - description: Process ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMRunResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: BPM процессын гүйлт эхлүүлэх
      tags:
      - bpm
  /bpm/tasks/{id}/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: Submission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.BPMSubmitTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.BPMRunResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Form даалгаврыг бөглөж дараагийн алхам руу шилжих
      tags:
      - bpm
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: /auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен
      (access token).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"